Rename files in common/LoginButton
